Move functions to commondatabuilder.
