+++ frontend/lib/tests/common-data.test.ts
@@ -1,16 +1,4 @@
-import { getDjangoChoiceLabel, validateDjangoChoices, allCapsToSlug, slugToAllCaps, filterDjangoChoices, DjangoChoices } from "../common-data";
-
-test('getDjangoChoiceLabel() works', () => {
-  expect(getDjangoChoiceLabel([['BLAH', 'boop']], 'BLAH')).toBe('boop');
-  expect(() => getDjangoChoiceLabel([['BLAH', 'boop']], 'NOPE'))
-    .toThrow('Unable to find label for value NOPE');
-});
-
-test("validateDjangoChoices() works", () => {
-  expect(validateDjangoChoices([['BLAH', 'boop']], ['BLAH'])).toBeUndefined();
-  expect(() => validateDjangoChoices([['BLAH', 'boop']], ['NOPE']))
-    .toThrow('Unable to find label for value NOPE');
-});
+import { allCapsToSlug, slugToAllCaps, toDjangoChoices } from "../common-data";
 
 test("allCapsToSlug() works", () => {
   expect(allCapsToSlug('BOOP_BLAP_BONK')).toBe('boop-blap-bonk');
@@ -20,8 +8,15 @@ test("slugToAllCaps() works", () => {
   expect(slugToAllCaps('boop-blap-bonk')).toBe('BOOP_BLAP_BONK');
 });
 
-test("filterDjangoChoices() works", () => {
-  const choices: DjangoChoices = [["FOO", "Foo"], ["BAR", "Bar"]];
-  expect(filterDjangoChoices(choices, ["FOO"])).toEqual([["BAR", "Bar"]]);
-  expect(filterDjangoChoices(choices, /^BA/)).toEqual([["FOO", "Foo"]]);
+test("toDjangoChoices() works", () => {
+  type Blarg = 'foo'|'bar';
+  const blargs: Blarg[] = ['foo', 'bar'];
+  const blargLabels = {
+    foo: 'hi',
+    bar: 'there'
+  };
+  expect(toDjangoChoices(blargs, blargLabels)).toEqual([
+    ['foo', 'hi'],
+    ['bar', 'there']
+  ]);
 });
