+++ frontend/lib/tests/signup-intent.test.tsx
@@ -1,17 +1,6 @@
-import { validateDjangoChoices } from "../common-data";
 import { OnboardingInfoSignupIntent } from "../queries/globalTypes";
 import { signupIntentFromOnboardingInfo, DEFAULT_SIGNUP_INTENT_CHOICE, getOnboardingRouteForIntent } from "../signup-intent";
 import { AppTesterPal } from "./app-tester-pal";
-
-type SignupIntentDjangoChoice = [OnboardingInfoSignupIntent, string];
-
-const SIGNUP_INTENT_CHOICES = require('../../../common-data/signup-intent-choices.json') as SignupIntentDjangoChoice[];
-
-test('OnboardingInfoSignupIntent has valid choices', () => {
-  for (let choice in OnboardingInfoSignupIntent) {
-    validateDjangoChoices(SIGNUP_INTENT_CHOICES, [choice, OnboardingInfoSignupIntent[choice]]);
-  }
-});
 
 test('signupIntentFromOnboardingInfo() works', () => {
   expect(signupIntentFromOnboardingInfo(null)).toStrictEqual(DEFAULT_SIGNUP_INTENT_CHOICE);
